renamed 'models' module to 'db'
