+++ securedrop_client/app.py
@@ -31,7 +31,7 @@ from securedrop_client import __version__
 from securedrop_client.logic import Client
 from securedrop_client.gui.main import Window
 from securedrop_client.resources import load_icon, load_css
-from securedrop_client.models import make_engine
+from securedrop_client.db import make_engine
 from securedrop_client.utils import safe_mkdir
 
 DEFAULT_SDC_HOME = '~/.securedrop_client'
