+++ securedrop_client/db.py
